+++ BetterBulldozer/BetterBulldozerMod.cs
@@ -5,12 +5,15 @@
 namespace Better_Bulldozer
 {
     using System;
+    using System.Collections.Generic;
     using System.IO;
+    using System.Linq;
     using System.Reflection;
     using Better_Bulldozer.Settings;
     using Better_Bulldozer.Systems;
     using Better_Bulldozer.Tools;
     using Colossal.IO.AssetDatabase;
+    using Colossal.Localization;
     using Colossal.Logging;
     using Game;
     using Game.Modding;
@@ -21,6 +24,7 @@ namespace Better_Bulldozer
     using System.Collections.Generic;
     using System.Linq;
     using Colossal;
+    using Colossal.Localization;
 #endif
 
     /// <summary>
@@ -32,12 +36,6 @@ namespace Better_Bulldozer
         /// A static ID for use with bindings.
         /// </summary>
         public static readonly string Id = "BetterBulldozer";
-
-        /// <summary>
-        /// Gets the version of the mod.
-        /// </summary>
-        internal string Version => Assembly.GetExecutingAssembly().GetName().Version.ToString(3);
-
         private Harmony m_Harmony;
 
         /// <summary>
@@ -48,6 +46,11 @@ namespace Better_Bulldozer
             get;
             private set;
         }
+
+        /// <summary>
+        /// Gets the version of the mod.
+        /// </summary>
+        internal string Version => Assembly.GetExecutingAssembly().GetName().Version.ToString(3);
 
         /// <summary>
         /// Gets ILog for mod.
@@ -92,6 +95,8 @@ namespace Better_Bulldozer
             m_Harmony.PatchAll();
             Logger.Info($"{nameof(BetterBulldozerMod)}.{nameof(OnLoad)} Loading en-US");
             GameManager.instance.localizationManager.AddSource("en-US", new LocaleEN(Settings));
+            Logger.Info($"[{nameof(BetterBulldozerMod)}] {nameof(OnLoad)} Loading localization for other languages.");
+            LoadNonEnglishLocalizations();
 #if DEBUG
             Logger.Info($"{nameof(BetterBulldozerMod)}.{nameof(OnLoad)} Exporting localization");
             var localeDict = new LocaleEN(Settings).ReadEntries(new List<IDictionaryEntryError>(), new Dictionary<string, int>()).ToDictionary(pair => pair.Key, pair => pair.Value);
@@ -134,6 +139,53 @@ namespace Better_Bulldozer
                 Settings = null;
             }
         }
+
+        private void LoadNonEnglishLocalizations()
+        {
+            Assembly thisAssembly = Assembly.GetExecutingAssembly();
+            string[] resourceNames = thisAssembly.GetManifestResourceNames();
+
+            try
+            {
+                Logger.Debug($"Reading localizations");
+
+                foreach (string localeID in GameManager.instance.localizationManager.GetSupportedLocales())
+                {
+                    string resourceName = $"{nameof(Better_Bulldozer)}.l10n.{localeID}.json";
+                    if (resourceNames.Contains(resourceName))
+                    {
+                        Logger.Debug($"Found localization file {resourceName}");
+                        try
+                        {
+                            Logger.Debug($"Reading embedded translation file {resourceName}");
+
+                            // Read embedded file.
+                            using StreamReader reader = new(thisAssembly.GetManifestResourceStream(resourceName));
+                            {
+                                string entireFile = reader.ReadToEnd();
+                                Colossal.Json.Variant varient = Colossal.Json.JSON.Load(entireFile);
+                                Dictionary<string, string> translations = varient.Make<Dictionary<string, string>>();
+                                GameManager.instance.localizationManager.AddSource(localeID, new MemorySource(translations));
+                            }
+                        }
+                        catch (Exception e)
+                        {
+                            // Don't let a single failure stop us.
+                            Logger.Error(e, $"Exception reading localization from embedded file {resourceName}");
+                        }
+                    }
+                    else
+                    {
+                        Logger.Debug($"Did not find localization file {resourceName}");
+                    }
+                }
+            }
+            catch (Exception e)
+            {
+                Logger.Error(e, "Exception reading embedded settings localization files");
+            }
+        }
+
 
     }
 }