+++ BetterBulldozer/ChangeLog.md
@@ -1,2 +1,4 @@
-﻿# Patch V1.3.10
-* Code Cleanup
+﻿# Patch V1.3.11
+* Re-did tool handling in a manner that works for v1.2.4
+* Removed I18n Everywhere dependency. 
+* Translations for officially supported languages are handled internally with embedded resource files.